refactor to left and right sidebar config files
