+++ src/configs/LeftSidebarConfig.tsx
@@ -3,13 +3,10 @@ import {
   BookOpen,
   Filter,
   Home,
-  List,
   Settings2,
   Share2,
 } from "lucide-react";
 import React from "react";
-import { ForceGraph3dLayoutMode } from "../AppConfig";
-import ForceGraphLayoutRadio from "../components/force-graph/ForceGraphLayoutRadio";
 import ForceGraphRenderConfigEditor from "../components/force-graph/ForceGraphRenderConfigEditor";
 import { CustomLayoutType } from "../core/layouts/CustomLayoutEngine";
 import { GraphologyLayoutType } from "../core/layouts/GraphologyLayoutEngine";
@@ -17,21 +14,6 @@ import { GraphvizLayoutType } from "../core/layouts/GraphvizLayoutEngine";
 import { PresetLayoutType } from "../core/layouts/LayoutEngine";
 import { extractPositionsFromNodes } from "../data/graphs/blobMesh";
 import styles from "../Sidebar.module.css";
-
-export interface SubMenuItem {
-  label: string;
-  onClick?: () => void;
-  content?: React.ReactNode;
-  customRender?: React.ReactNode;
-}
-
-export interface MenuItem {
-  id: string;
-  icon: React.ReactNode;
-  label: string;
-  content?: React.ReactNode;
-  subMenus?: SubMenuItem[];
-}
 
 const allLayoutLabels = [
   ...Object.values(GraphvizLayoutType),
@@ -228,42 +210,7 @@ export const createDefaultLeftMenus = ({
   },
 ];
 
-export const createDefaultRightMenus = (
-  renderLegends: () => React.ReactNode,
-  isForceGraph3dActive: boolean,
-  forceGraphLayout: string,
-  onForceGraphLayoutChange: (layout: string) => void,
-  isDarkMode: boolean
-) => [
-  {
-    id: "legends",
-    icon: <List size={20} className={styles.menuIcon} />,
-    label: "Legends",
-    content: (
-      <div className={styles.optionsPanelContainer}>{renderLegends()}</div>
-    ),
-  },
-  {
-    id: "settings",
-    hideHeader: true, // Add this flag to hide the header
-    icon: <Settings2 size={20} className={styles.menuIcon} />,
-    label: "Settings",
-    alwaysExpanded: true, // Add this flag to keep it expanded
-    content: (
-      <div className={styles.optionsPanelContainer}>
-        {isForceGraph3dActive && (
-          <ForceGraphLayoutRadio
-            layout={forceGraphLayout as ForceGraph3dLayoutMode}
-            onLayoutChange={onForceGraphLayoutChange}
-            isDarkMode={isDarkMode}
-          />
-        )}
-      </div>
-    ),
-  },
-];
-
-export const footerContent = (isOpen: boolean) => (
+export const leftFooterContent = (isOpen: boolean) => (
   <a
     href="https://aesgraph.github.io/unigraph/"
     target="_blank"
